// helpers/types.ts
export enum eMantleNetwork{
  mantleTestnet = "mantleTestnet",
  mantle = "mantle"
}
// ...
export interface IBaseConfiguration {
  SynthexAddress?: tEthereumAddress;
  MarketId: string;
  ATokenNamePrefix: string;
  StableDebtTokenNamePrefix: string;
  VariableDebtTokenNamePrefix: string;
  SymbolPrefix: string;
  ProviderId: number;
  PythId?: string;
  TestnetMarket?: boolean;
  ProviderRegistryOwner?: iParamsPerNetwork<tEthereumAddress | undefined>;

  FallbackOracle?: OracleConfig;
  PriceOracle: OracleConfig;

  WrappedTokenGateway?: iParamsPerNetwork<tEthereumAddress>;
  ReserveFactorTreasuryAddress: iParamsPerNetwork<tEthereumAddress>;
  StableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  VariableDebtTokenImplementation?: iParamsPerNetwork<tEthereumAddress>;
  ReserveAssets?: iParamsPerNetwork<SymbolMap<tEthereumAddress>>;
  
  ReservesConfig: SymbolMap<IReserveParams>;
  WrappedNativeTokenSymbol: string; // ex: WETH or WMATIC
  IncentivesConfig: IncentivesConfig;
  EModes: SymbolMap<EMode>;
  L2PoolEnabled?: iParamsPerNetwork<boolean>;
  StkAaveProxy?: iParamsPerNetwork<tEthereumAddress>;
  ParaswapRegistry?: iParamsPerNetwork<tEthereumAddress>;
  FlashLoanPremiums: {
    total: number;
    protocol: number;
  };
  RateStrategies: IStrategy;
}

export interface ICommonConfiguration extends IBaseConfiguration {}

export interface IMosaicConfiguration extends ICommonConfiguration {
  // ReservesConfig: iMosaicPoolAssets<IReserveParams>;
}
// ...
export enum AssetType {
  AToken,
  VariableDebtToken,
  StableDebtToken,
}

export enum TransferStrategy {
  PullRewardsStrategy,
  StakedRewardsStrategy,
}